requirements.txt
python-dotenv
openai
piexif==1.1.3
django_celery_results
sitemap
apscheduler
django-user-agents==0.4.0
httpx
django-cors-headers
django-celery-beat
django-cron
django-crontab==0.7.1
google-search-results
django-environ
httplib2
oauth2client
gunicorn
whitenoise
google-api-python-client
pandas
ultimate-sitemap-parser
django-ckeditor
django-redis
celery-redbeat
django-dbbackup
django-jazzmin
cloudflare
python-whois
transformers
# sentencepiece
nltk
numpy
scipy
scikit-learn
networkx
torch
google-ads
amqp==5.1.1
    # via kombu
anyascii==0.3.1
    # via wagtail
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
beautifulsoup4==4.9.3
    # via wagtail
billiard==3.6.4.0
    # via celery
celery[redis]==5.2.7
    # via -r .\requirements\requirements.in
celery-progress==0.1.2
    # via -r .\requirements\requirements.in
certifi==2022.5.18.1
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colorama==0.4.5
    # via click
coreapi==2.3.3
    # via -r .\requirements\requirements.in
coreschema==0.0.4
    # via coreapi
cryptography==37.0.2
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
deprecated==1.2.13
    # via redis
dj-stripe==2.6.1
    # via -r .\requirements\requirements.in
django==3.2.13
    # via
    #   -r .\requirements\requirements.in
    #   dj-stripe
    #   django-allauth
    #   django-anymail
    #   django-filter
    #   django-modelcluster
    #   django-permissionedforms
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   jsonfield
    #   wagtail
django-allauth==0.50.0
    # via -r .\requirements\requirements.in
django-anymail[mailgun]==8.6
    # via -r .\requirements\requirements.in
django-filter==21.1
    # via wagtail
django-modelcluster==6.0
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-taggit==2.1.0
    # via wagtail
django-treebeard==4.5.1
    # via wagtail
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangorestframework-datatables==0.7.0
    # via
    #   -r .\requirements\requirements.in
    #   wagtail
draftjs-exporter==2.1.7
    # via wagtail
et-xmlfile==1.1.0
    # via openpyxl
html5lib==1.1
    # via wagtail
httplib2==0.20.4
    # via
    #   -r .\requirements\requirements.in
    #   oauth2client
idna==3.3
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jsonfield==3.1.0
    # via dj-stripe
kombu==5.2.4
    # via celery
l18n==2021.3
    # via wagtail
markupsafe==2.1.1
    # via jinja2
oauth2client==4.1.3
    # via -r .\requirements\requirements.in
oauthlib==3.2.0
    # via requests-oauthlib
openpyxl==3.0.10
    # via tablib
packaging==21.3
    # via redis
pillow==9.1.1
    # via wagtail
prompt-toolkit==3.0.29
    # via click-repl
psycopg2-binary==2.9.3
    # via -r .\requirements\requirements.in
pyasn1==0.4.8
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via oauth2client
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.4.0
    # via django-allauth
pyparsing==3.0.9
    # via
    #   httplib2
    #   packaging
python3-openid==3.2.0
    # via django-allauth
pytz==2022.1
    # via
    #   celery
    #   django
    #   django-modelcluster
    #   djangorestframework
    #   l18n
redis==4.3.1
    # via celery
requests==2.27.1
    # via
    #   coreapi
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
    #   stripe
    #   wagtail
requests-oauthlib==1.3.1
    # via django-allauth
rsa==4.8
    # via oauth2client
six==1.16.0
    # via
    #   click-repl
    #   html5lib
    #   l18n
    #   oauth2client
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
stripe==3.2.0
    # via
    #   -r .\requirements\requirements.in
    #   dj-stripe
tablib[xls,xlsx]==3.2.1
    # via wagtail
telepath==0.2
    # via wagtail
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.9
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wagtail==3.0
    # via -r .\requirements\requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
willow==1.4.1
    # via wagtail
wrapt==1.14.1
    # via deprecated
xlrd==2.0.1
    # via tablib
xlsxwriter==3.0.3
    # via wagtail
xlwt==1.3.0
    # via tablib